Small updates.
A* seems to be almost working now. I still want to change the way i keep track of nodes to increase the speed. Also noticed I didn't use camelcase a lot of places so thats on the todo list too. Finally some stuff needs rewriting, like the rendering parts and the direction functions of the snake.