Algortihm is able to somewhat play snake.
The a* algo now seems to somewhat be able to play the game. Edges seem to screw up the path finding. Especially when getting food right on the border. This might be speed of the a* not being fast enough to keep up with the interval at which the snake moves. It does however sometimes move towards the border even though it wasn't on his path originally so something still is not working 100%.

Also changed how to make the snake move direction. All directions are now separate functions. By keeping track of the direction it had at it's last move I've been able to stop the instant turn around thing when being a single block. For example, if you were heading north, quickly pressing left or right and then down within the same interval would cause the snake to u-turn. This is now impossible.