Fixed path finder.
By rendering the path I was able to conclude that the snake took the path in reverse. This caused a lot of strange behaviour, calling a simple reverse fixed this. 

Rendering still needs to be made nicer. It works for developing now but needs to made nicer, also to show where a the body parts connect. Currently I fill the whole square out of ease but this makes it impossible to see what part is the head or where the connection between parts are when it has moved itself in, for example, a square.

First I want to move everything to an actual grid, this 2d array should speed everything up since I won't have to do so many for loops looping through whole arrays.